Move csv reader to io package along with any vestigial remains
