--- pdtable/writers/__init__.py
@@ -1,1 +1,0 @@
-from ._csv import write_csv
